fix return type in last commit for AutomationLine::control_point_box_size(); catch attempts to use new session dialog to open existing sessions
git-svn-id: svn://localhost/ardour2/trunk@1042 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_line.h
@@ -217,7 +217,7 @@ class AutomationLine : public sigc::trackable, public PBD::StatefulThingWithGoin
 	void reset_line_coords (ControlPoint&);
 	void update_line ();
 
-	uint32_t control_point_box_size ();
+	double control_point_box_size ();
 
 	struct ModelRepresentation {
 	    ARDOUR::AutomationList::iterator start;
